Introduce new DateRange class
This is a value object [1] representing a range of dates. It is instantiated
using a `begins_on` date and an `ends_on` date. If `begins_on` is not specified
then the range starts from a negatively infinite date, and if `ends_on` is not
specified then the range ends on a positively infinite date. The start and end
dates are converted to a date (in case they aren't already).

Since it's a value object, I've implemented equality-related methods so that
`DateRange`'s with the same start and end dates are equal/equivalent to each
other. These implementations are based on those in the `values` gem [2].

I'm planning to use this in the new calculate-your-part-year-profit flow, but I
expect it could also be used in some of the existing flows.

[1]: http://c2.com/cgi/wiki?ValueObject
[2]: https://rubygems.org/gems/values