Respect user provided input for paternity leave duration
Bug report: Customer feedback has shown that the paternity calculator on GOV.UK
is producing incorrect results. If an employee requests 1 week paternity pay
(meeting the relevant conditions) the calculator should give an entitlement of
1 week at £139.58 (or 90% of AWE if lower) but it is producing an entitlement
of £279.16 which is equivalent to 2 weeks paternity pay. Also within the
calculation the information states the latest date to claim SPP is 28 days
before they want it to start, this is incorrect as of April 2015 and should be
15 weeks before they want it to start.

Example path: /maternity-paternity-calculator/y/paternity/no/2015-10-01/2015-10-01/yes/yes/yes/yes/yes/2015-10-01/one_week/2015-06-19/2015-04-24/weekly/5000.0/weekly_starting

The calculator was ignoring user input for paternity leave duration when
working out payment amounts. It was only using it for leave end date
calculations implemented in the flow file. This change sets the
paternity_leave_duration value and uses it in the calculator + reuses the
existing MaternityPaternityCalculator#pay_end_date method to calculate the
leave_end_date value.