Move usual_work_days? question up in the calculate SSP flow
It has been reported by HMRC that currently this Smart Answer does not
calculate the number of sick days correctly e.g. when calculating whether the
employee has been sick for the minimum of 3 days it does not take into account
the usual work days. In order to fix this we need to move up the question that
collects employee's working patterns.

The usual_work_days? question used to be the last one in the flow and had the
logic for deciding the next outcome node (out of 4 possible). Three questions
were leading to the usual_work_days? question. Removing it from the last spot
in the flow means those three questions will need to use the logic to decide
the outcome node.

In order to avoid duplicating this non-trivial logic I've extracted it to the
private OutcomeDecision class (see earlier commit).