Move calculate blocks in Q to precalculate blocks in outcome
We want to move the `:usual_work_days?` question much earlier in the flow and
so we need to move these `calculate` blocks to the relevant outcome and convert
them into `precalculate` blocks so they are available in the outcome template.

Note that the block for `formatted_sick_pay_weekly_amounts` was instantiating
its own copy of the calculator, but since we no longer have easy access to the
`response` in the `precalculate` block and we already have access to an instance
of the calculator, it's simpler to use that.