Remove analytics:analyse rake task
This task hasn't been working since we started work on the project about five
months ago - see [1] - and nobody has complained about it.

We know the analytics folk are able to do some analysis of Smart Answers
directly in Google Analytics by making use of the functionality we added in [2]
and other related PRs.

So this code is just cluttering up the codebase unnecessarily and is safe to
remove.

[1]: #1670
[2]: #1907